Explicitly disignate the size of type in kernel codes.

+++ utils/kernel.py
@@ -129,25 +129,21 @@ void step_minimization(
 ''', 'step_minimization')
 
 step_minimization_integer = cp.RawKernel(r'''
-inline __device__ long long int __ull_as_ll(
-    unsigned long long int val
-) {
-    return *(long long int*)&val;
+inline __device__ int64_t __ull_as_ll(uint64_t val) {
+    return *(int64_t*)&val;
 }
 
-inline __device__ unsigned long long int __ll_as_ull(
-    long long int val
-) {
-    return *(unsigned long long int*)&val;
+inline __device__ uint64_t __ll_as_ull(int64_t val) {
+    return *(uint64_t*)&val;
 }
 
-inline __device__ long long int atomicAdd(
-    long long int* addr, 
-    long long int val
+inline __device__ int64_t atomicAdd(
+    int64_t* addr, 
+    int64_t val
 ) {
-    unsigned long long int* addr_as_ull = (unsigned long long int*)addr;
-    unsigned long long int old = *addr_as_ull;
-    unsigned long long int assumed;
+    uint64_t* addr_as_ull = (uint64_t*)addr;
+    uint64_t old = *addr_as_ull;
+    uint64_t assumed;
     do {
         assumed = old;
         old = atomicCAS(
@@ -161,41 +157,41 @@ inline __device__ long long int atomicAdd(
 
 extern "C" __global__
 void step_minimization_integer(
-    const int* LivImg, 
-    const int* RefImg, 
-    int width, 
-    int height, 
+    const int32_t* LivImg, 
+    const int32_t* RefImg, 
+    int32_t width, 
+    int32_t height, 
     const float* WarpMat, 
     float* GenMat, 
-    int dim_param, 
-    long long int* MatA, 
-    long long int* Matb, 
-    long long int* Error
+    int32_t dim_param, 
+    int64_t* MatA, 
+    int64_t* Matb, 
+    int64_t* Error
 ) {
-    const int tx = blockIdx.x*blockDim.x + threadIdx.x;
-    const int ty = blockIdx.y*blockDim.y + threadIdx.y;
+    const int32_t tx = blockIdx.x*blockDim.x + threadIdx.x;
+    const int32_t ty = blockIdx.y*blockDim.y + threadIdx.y;
 
     if(tx >= width || ty >= height) { return; }
 
-    const float wz =  WarpMat[6] * tx + WarpMat[7] * ty + WarpMat[8];
-    const int   wx = (WarpMat[0] * tx + WarpMat[1] * ty + WarpMat[2]) / wz;
-    const int   wy = (WarpMat[3] * tx + WarpMat[4] * ty + WarpMat[5]) / wz;
+    const float   wz =  WarpMat[6] * tx + WarpMat[7] * ty + WarpMat[8];
+    const int32_t wx = (WarpMat[0] * tx + WarpMat[1] * ty + WarpMat[2]) / wz;
+    const int32_t wy = (WarpMat[3] * tx + WarpMat[4] * ty + WarpMat[5]) / wz;
 
     if(wx < 0 || wx >= width || wy < 0 || wy >= height) { return; }
 
-    int diff = LivImg[wx + wy * width] - RefImg[tx + ty * width];
-    long long int err = diff * diff;
+    int32_t diff = LivImg[wx + wy * width] - RefImg[tx + ty * width];
+    int64_t err = diff * diff;
     atomicAdd(Error, err);
 
-    const long long int scale = 1.0;
-    int Ix = RefImg[min(tx + 1, width - 1) + ty * width] - RefImg[max(tx - 1, 0) + ty * width];
-    int Iy = RefImg[tx + min(ty + 1, height -1) * width] - RefImg[tx + max(ty - 1, 0) * width];
-    int Iz = (tx * Ix + ty * Iy) * scale;
+    const int64_t scale = 1.0;
+    int32_t Ix = RefImg[min(tx + 1, width - 1) + ty * width]-RefImg[max(tx - 1, 0) + ty * width];
+    int32_t Iy = RefImg[tx + min(ty + 1, height -1) * width]-RefImg[tx + max(ty - 1, 0) * width];
+    int32_t Iz = (tx * Ix + ty * Iy) * scale;
 
-    int i, j;
+    int32_t i, j;
     float* gen_mat;
-    long long int elm_A, elm_b;
-    long long int param[9]; // dim_param <= 9
+    int64_t elm_A, elm_b;
+    int64_t param[9]; // dim_param <= 9
     for(i = 0; i < dim_param; i++) {
         gen_mat = GenMat + 9 * i;
         param[i] =  - Ix * (gen_mat[0] * tx + gen_mat[1] * ty + gen_mat[2])
